moved init and initpop to default_fitters (from default initializers). Fixed bug in TF nmse calc. Made per_cell=True default. Added a bunch of TF options to init.tf
